requirements.txt
ipykernel
pandas
openpyxl>=3.0.0
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost
pymongo            # it help to connect mongodb atlas
from_root          # finding all root folders
evidently==0.2.8   # used for data drift predictions
dill
PyYAML
neuro_mf
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
mypy 
pylint
xlsxwriter>=1.3.0
logging
rich

-e .                 